+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- better macOS support with more event properties [#95](https://github.com/PostHog/posthog-ios/pull/95)
+- better macOS support with more event properties [#96](https://github.com/PostHog/posthog-ios/pull/96)
 
 ## 3.0.0-beta.3 - 2024-01-11